requirements.txt
requests==2.31.0
anthropic==0.18.1
tabulate==0.9.0
python-dotenv==1.0.0 